correction du chronometre
git-svn-id: https://dev.isen.fr/svn/2012-2013/cir3/java/project/group5/geneticAlgo/GeneticAlgorithm@126 84736ae8-fca8-4d2c-b905-760bd7953024

+++ src/main/java/geneticalgorithm/Problems/Problem.java
@@ -195,7 +195,7 @@ public abstract class Problem extends Model {
         return this.stopCriteria.getEvolutionCriterion();
     }
     
-    public boolean stopCriteriaAreReached(int stepCount, int time, double evolutionCoeff){
+    public boolean stopCriteriaAreReached(int stepCount, long time, double evolutionCoeff){
         return this.stopCriteria.areReached(stepCount, time, evolutionCoeff);
     }
 }
